Updating console apps to .NET 8 and WPF examples to .NET 4.8.1
