+++ src/CsvConverter.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.5.33627.172
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CsvConverter", "CsvConverter\CsvConverter.csproj", "{FB7F636B-DC53-45BC-A528-8C86C952BE1B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleCoreExample1", "Examples\CsvConverter.SimpleCoreExample1\SimpleCoreExample1.csproj", "{6D022FAC-E925-41E4-9421-397C8E4A5A52}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{33FABC1D-20AA-4F37-AE6B-33CABE5783AC}"
 	ProjectSection(SolutionItems) = preProject
 		Examples\Readme.md = Examples\Readme.md
@@ -24,11 +22,13 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleDotNetExample1", "Exa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleDotNetExample2", "Examples\CsvConverter.SimpleDotNetExample2\SimpleDotNetExample2.csproj", "{A3327692-A117-404B-B399-807CFCB2F56B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleCoreExample2", "Examples\CsvConverter.SimpleCoreExample2\SimpleCoreExample2.csproj", "{7A778F22-119B-4181-B908-E20A4CA80085}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CsvConverter.Core.IntegrationTests", "CsvConverter.Core.IntegrationTests\CsvConverter.Core.IntegrationTests.csproj", "{A75C2426-B04A-425C-B72E-537AD89F60BB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CsvConverter.Core.Tests", "CsvConverter.Core.Tests\CsvConverter.Core.Tests.csproj", "{C319595F-22B0-4536-A456-36B4A4CACA31}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleDotNetConsoleExample1", "Examples\CsvConverter.SimpleDotNetConsoleExample1\SimpleDotNetConsoleExample1.csproj", "{C854B988-711E-4ECF-A0E2-A69AB65FD577}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleDotNetConsoleExample2", "Examples\CsvConverter.SimpleDotNetConsoleExample2\SimpleDotNetConsoleExample2.csproj", "{7FD12A82-6134-4C04-80BE-E7D29FB9960A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -46,14 +46,6 @@ Global
 		{FB7F636B-DC53-45BC-A528-8C86C952BE1B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FB7F636B-DC53-45BC-A528-8C86C952BE1B}.Release|x86.ActiveCfg = Release|Any CPU
 		{FB7F636B-DC53-45BC-A528-8C86C952BE1B}.Release|x86.Build.0 = Release|Any CPU
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52}.Debug|x86.Build.0 = Debug|Any CPU
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52}.Release|x86.ActiveCfg = Release|Any CPU
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52}.Release|x86.Build.0 = Release|Any CPU
 		{9440032C-AE95-4BAC-9764-D5F07711C815}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{9440032C-AE95-4BAC-9764-D5F07711C815}.Debug|Any CPU.Build.0 = Debug|x86
 		{9440032C-AE95-4BAC-9764-D5F07711C815}.Debug|x86.ActiveCfg = Debug|x86
@@ -89,14 +81,6 @@ Global
 		{A3327692-A117-404B-B399-807CFCB2F56B}.Release|Any CPU.ActiveCfg = Release|x86
 		{A3327692-A117-404B-B399-807CFCB2F56B}.Release|x86.ActiveCfg = Release|x86
 		{A3327692-A117-404B-B399-807CFCB2F56B}.Release|x86.Build.0 = Release|x86
-		{7A778F22-119B-4181-B908-E20A4CA80085}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7A778F22-119B-4181-B908-E20A4CA80085}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7A778F22-119B-4181-B908-E20A4CA80085}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7A778F22-119B-4181-B908-E20A4CA80085}.Debug|x86.Build.0 = Debug|Any CPU
-		{7A778F22-119B-4181-B908-E20A4CA80085}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7A778F22-119B-4181-B908-E20A4CA80085}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7A778F22-119B-4181-B908-E20A4CA80085}.Release|x86.ActiveCfg = Release|Any CPU
-		{7A778F22-119B-4181-B908-E20A4CA80085}.Release|x86.Build.0 = Release|Any CPU
 		{A75C2426-B04A-425C-B72E-537AD89F60BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A75C2426-B04A-425C-B72E-537AD89F60BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A75C2426-B04A-425C-B72E-537AD89F60BB}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -113,20 +97,36 @@ Global
 		{C319595F-22B0-4536-A456-36B4A4CACA31}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C319595F-22B0-4536-A456-36B4A4CACA31}.Release|x86.ActiveCfg = Release|Any CPU
 		{C319595F-22B0-4536-A456-36B4A4CACA31}.Release|x86.Build.0 = Release|Any CPU
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577}.Debug|x86.Build.0 = Debug|Any CPU
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577}.Release|x86.ActiveCfg = Release|Any CPU
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577}.Release|x86.Build.0 = Release|Any CPU
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A}.Debug|x86.Build.0 = Debug|Any CPU
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A}.Release|x86.ActiveCfg = Release|Any CPU
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{6D022FAC-E925-41E4-9421-397C8E4A5A52} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
 		{9440032C-AE95-4BAC-9764-D5F07711C815} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
 		{4C433F80-2987-4F0F-A27F-84615F0CF33B} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
 		{9EF7291B-3373-41F9-B657-784121D391ED} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
 		{9BF1F121-B186-458F-A8F1-C21E9CFFACE6} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
 		{A3327692-A117-404B-B399-807CFCB2F56B} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
-		{7A778F22-119B-4181-B908-E20A4CA80085} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
 		{A75C2426-B04A-425C-B72E-537AD89F60BB} = {902950D2-7DFF-4A1E-A481-9180D990ADEA}
 		{C319595F-22B0-4536-A456-36B4A4CACA31} = {902950D2-7DFF-4A1E-A481-9180D990ADEA}
+		{C854B988-711E-4ECF-A0E2-A69AB65FD577} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
+		{7FD12A82-6134-4C04-80BE-E7D29FB9960A} = {33FABC1D-20AA-4F37-AE6B-33CABE5783AC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {EBED656A-5AC0-4B1F-9E0C-2E9BC3BF9E96}
