--- src/Examples/CsvConverter.SimpleCoreExample1/Model/TestData.cs
@@ -1,19 +1,0 @@
-﻿using CsvConverter;
-
-namespace SimpleCoreExample1
-{
-    public class TestData
-    {
-        [CsvConverter(ColumnIndex = 1)]
-        public string FieldName { get; set; }
-
-        [CsvConverter(ColumnIndex = 2)]
-        public string FieldDescription { get; set; }
-
-        [CsvConverter(ColumnIndex = 3)]
-        public string StartPosition { get; set; }
-
-        [CsvConverter(ColumnIndex = 4)]
-        public string EndPosition { get; set; }
-    }
-}
