--- src/Examples/CsvConverter.SimpleCoreExample1/Readme.md
@@ -1,12 +1,0 @@
-﻿# Summary
-This is a console
-
-This examples shows
-1. How to create a file 
-2. How to read from a file
-3. How to create a custom converter.  In this case it reads a number into an enum and writes an enum as a number.
-
-# Running
-1. To run it, right click on the project and select "Set as StartUp Project".
-2. Press F5.
-
