+++ src/Examples/CsvConverter.SimpleDotNetConsoleExample1/Readme.md
@@ -1,0 +1,12 @@
+﻿# Summary
+This is a console
+
+This examples shows
+1. How to create a file 
+2. How to read from a file 
+
+# Running
+1. Create a c:\temp directory.
+2. To run it, right click on the project and select "Set as StartUp Project".
+3. Press F5.
+
