+++ src/Examples/CsvConverter.SimpleDotNetConsoleExample2/Readme.md
@@ -7,6 +7,7 @@ This examples shows
 3. How to use a couple of the built in pre-converters.
 
 # Running
-1. To run it, right click on the project and select "Set as StartUp Project".
-2. Press F5.
+1. Create a c:\temp directory.
+2. To run it, right click on the project and select "Set as StartUp Project".
+3. Press F5.
