+++ src/Examples/Readme.md
@@ -1,10 +1,13 @@
 ﻿Examples
 
 - Simple Examples
-	- SimpleCoreExample1 (.NET Core example)
+	- SimpleDotNetConsoleExample1 (.NET Console example)
+		1. How to create a file 
+		1. How to read from a file 
+    - SimpleDotNetConsoleExample2 (.NET Console example)
 		1. How to create a file 
 		1. How to read from a file
-		1. How to create a custom converter.  In this case it reads a number into an enum and writes an enum as a number.
+		1. How to use a couple of the built in pre-converters.	
 	- SimpleDotNetExample1 (WPF Example)
 		1. How to create a file 
 		1. How to read from a file that has a header row.
